refactor(grid): move compute_latlon_grid
Move compute_latlon_grid function from grid_helper module to grid_goesdr
module, and update imports.
